requirements.txt
grpcio
grpcio-tools
matplotlib
networkx
numpy
pandas
protobuf
scikit-learn
sicpy
